sign conversion fix

+++ package/prefixdb/prefixdb/multidb/wrocksdb/wrocksdb.cpp
@@ -719,7 +719,7 @@ bool wrocksdb::backup()
   PREFIXDB_LOG_BEGIN("CreateNewBackup...")
   if (auto db = _db)
   {
-    int progress = 0;
+    size_t progress = 0;
     status = _backup->CreateNewBackup( db.get(), true, 
       [progress]() mutable { PREFIXDB_LOG_PROGRESS("CreateNewBackup...." << std::string(progress++, '.') ) });
   }
